update for new deployment

+++ scripts/tmux-run-thing.sh
@@ -16,17 +16,20 @@ PRO_DIR='$HOME/projects/hongbao-ms'
 
 tmux has-session -t $SESSION_NAME 2>/dev/null
 if [ $? = 0 ]; then
-  tmux kill-session -t $SESSION_NAME
+  PANE_NUM=$(tmux list-panes -t $SESSION_NAME | wc -l)
+  if [ $PANE_NUM -lt $NUM ]; then
+    tmux kill-session -t $SESSION_NAME
+    tmux new-session -s $SESSION_NAME -d
+    for (( i=1; i<$NUM; i++ ))
+    do
+      tmux split-window -t $SESSION_NAME:0
+    done
+    tmux select-layout -t $SESSION_NAME:0 tiled
+  fi
 fi
-
-tmux new-session -s $SESSION_NAME -d
-for (( i=1; i<$NUM; i++ ))
-do
-  tmux split-window -t $SESSION_NAME:0
-done
-tmux select-layout -t $SESSION_NAME:0 tiled
 
 for (( i=0; i<$NUM; i++))
 do
+  tmux send-keys -t $SESSION_NAME:0.$i C-c
   tmux send-keys -t $SESSION_NAME:0.$i "bash $PRO_DIR/scripts/docker-run-thing.sh $CONFIG $NODE $i" C-m
 done
